arch64: mm: ioremap: add casts to avoid gcc9 warnings

+++ arch/arm64/mm/ioremap.c
@@ -93,10 +93,10 @@ int pci_ioremap_io(unsigned int offset, phys_addr_t phys_addr)
 {
 	BUG_ON(offset + SZ_64K > IO_SPACE_LIMIT);
 
-	return ioremap_page_range(PCI_IOBASE + offset,
-				  PCI_IOBASE + offset + SZ_64K,
-				  phys_addr,
-				  __pgprot(PROT_DEVICE_nGnRE));
+	return ioremap_page_range((unsigned long)PCI_IOBASE + offset,
+				 (unsigned long)PCI_IOBASE + offset + SZ_64K,
+				 phys_addr,
+				 __pgprot(PROT_DEVICE_nGnRE));
 }
 EXPORT_SYMBOL_GPL(pci_ioremap_io);
 #endif
